test_combine.py (#32253)
